Sample random weights on edges of a binary Stochastic Block Model

This function draws weights according to a user-defined distribution for the edges of an existing binary SBM.

rWeightSBM(anSBM, family = c("gaussian", "poisson"), theta)

Arguments

anSBM
SBM sampled from the rSBM function with nbBlock

family
character describing the distribution used for the weigths

theta
list embedding parameters required for the distribution of the weights. See details.

Elements in the theta should be named as follows, depending on the argument family, with dimension matching the number of blocks in the original SBM:

Gaussiantheta$mu, a (nbBlock x nbBlock) matrix of means; theta$sigma, a (nbBlock x nbBlock) matrix of standa deviation

Poissontheta$lambda, , a (nbBlock x nbBlock) matrix of means.

Otherszoon available.

Value

an SBM weight weigthed edges

Examples

## graph parameters
nbNodes  <- 90
blockProp <- c(.5, .25, .25)   # group proportions
nbBlock   <- length(blockProp) # number of blocks
connectParam <- diag(.4, nbBlock) + 0.05 # connectivity matrix: affiliation network

## Graph Sampling
mySBM <- rSBM(nbNodes, connectParam, blockProp)

## Sampling Gaussian weights
mu_within   <- 4 ; sigma_within  <- .5
mu_between  <- 2 ; sigma_between <- .5
theta <- list()
theta$mu    <- matrix(mu_between   , nbBlock, nbBlock); diag(theta$mu)    <- mu_within    # means
theta$sigma <- matrix(sigma_between, nbBlock, nbBlock); diag(theta$sigma) <- sigma_within # sd

mySBM <- rWeightSBM(mySBM, "gaussian", theta)

#> Error in igraph::gsize(mySBM): object 'mySBM' not found

hist(igraph::E(mySBM)$weight,  breaks = sqrt(igraph::gsize(mySBM)))

#> Error in hist.default(igraph::E(mySBM)$weight, breaks = sqrt(igraph::gsize(mySBM))): 'x' must be numeric